api split threading
Move ds specific classes and functions to ds and make public when needed
