+++ qcodes/utils/__init__.py
@@ -20,12 +20,14 @@ from .json_utils import NumpyJSONEncoder
 from .partial_utils import partial_with_docstring
 from .path_helpers import get_qcodes_path, get_qcodes_user_path
 from .spyder_utils import add_to_spyder_UMR_excludelist
+from .threading_utils import RespondingThread, thread_map
 
 __all__ = [
     "DelayedKeyboardInterrupt",
     "DelegateAttributes",
     "NumpyJSONEncoder",
     "QCoDeSDeprecationWarning",
+    "RespondingThread",
     "attribute_set_to",
     "checked_getattr",
     "convert_legacy_version_to_supported_version",
@@ -41,4 +43,5 @@ __all__ = [
     "partial_with_docstring",
     "qcodes_abstractmethod",
     "strip_attrs",
+    "thread_map",
 ]
